Don't depend on Prawn to provide prawn.png
This seems very brittle and unnessesary. Especially for the specs; it doesn't matter what image is used here, just that *any* image is used. For the manual, it could be argued that we get it from Prawn. But I think it makes our code too brittle for too little gain. Better to have evrything self-contained.

Fixes all Travis errors.