Import the Prawn images needed for tests and manual
Since the Prawn team, due Gem's best practices of not including nothing
that it's not required at runtime.
An issue was opened in the project at
https://github.com/prawnpdf/prawn/issues/1041 to ask for clarification
on this and as result we should have what we need in our repo and not
relying other gems.

So in this commit we've added the needed images and replaced their use
in code and manual to make use of the local ones.