Fix build setup and pivot to GitHub Actions
This fixes unit tests which failes because they tried to use
an image files from another gem and did not find it in the
expected location. We simply add the file as a fixture
to the project at hand.

Also, this project was still using Travis CI **dot org**
which was deprecated in favor of **dot com**, and then again
Travis CI introduced restrictions to protect their
infrastructure which are just detremental to OSS development:

https://www.jeffgeerling.com/blog/2020/travis-cis-new-pricing-plan-threw-wrench-my-open-source-works

Hence we move all tests over to GitHub Actions.